lumera/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
opencv-python==4.10.0.84
numpy==1.26.4
scikit-learn==1.5.2
tensorflow==2.17.0
keras==3.4.1
matplotlib==3.9.0
kagglehub==0.2.9